Add ID.me OAuth, Netlify deployment, and refactor server structure
