+++ .env.example
@@ -12,3 +12,12 @@ SMTP_PASS=your_smtp_password
 FROM_EMAIL=careers@corecrewlogistics.com
 TO_EMAIL=corecrewlogistics@gmail.com
 
+# ID.me OpenID Connect
+# Production: https://api.id.me
+# Sandbox: https://api.idmelabs.com
+IDME_ISSUER=https://api.idmelabs.com
+IDME_CLIENT_ID=your_client_id
+IDME_CLIENT_SECRET=your_client_secret
+IDME_REDIRECT_URI=http://localhost:3000/auth/idme/callback
+IDME_SCOPE=openid email profile
+
