+++ netlify/functions/server.js
@@ -1,0 +1,5 @@
+const serverless = require('serverless-http');
+const app = require('../../src/app');
+
+module.exports.handler = serverless(app);
+
